# doc/api-documentation.yml
openapi: 3.0.3
info:
  title: AurAchieve Backend API
  description: The official documentation for AurAchieve's backend API, which powers the AurAchieve mobile app.
  version: 1.0.0
servers:
  - url: https://auraascend-fgf4aqf5gubgacb3.centralindia-01.azurewebsites.net
paths:
  /api/user/profile:
    get:
      summary: Get user profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Profile fetched successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: string
                  name:
                    type: string
                  email:
                    type: string
                  aura:
                    type: integer
                  validationCount:
                    type: integer
                  lastValidationResetDate:
                    type: string
        '500':
          description: Error fetching user profile
  /api/tasks:
    get:
      summary: Get all tasks for user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Successfully fetched all the tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '500':
          description: Failed to fetch tasks
    post:
      summary: Create a new task
      description: This endpoint creates a new task for the user.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                taskCategory:
                  type: string
                  enum: [normal, timed]
                durationMinutes:
                  type: integer
      responses:
        '201':
          description: Task created
        '400':
          description: Validation error
        '500':
          description: Internal error
  /api/tasks/{taskId}/complete:
    post:
      summary: Complete a normal image-verifiable task
      description: Marks a normal task as verified.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                imageBase64:
                  type: string
      responses:
        '200':
          description: Task completed
        '400':
          description: Validation error
        '404':
          description: Task not found
        '500':
          description: Internal error
  /api/tasks/{taskId}/complete-normal-non-verifiable:
    post:
      summary: Complete a normal non-image-verifiable task
      description: Marks a normal non-image-verifiable task as completed.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task completed
        '400':
          description: Validation error
        '404':
          description: Task not found
        '500':
          description: Internal error
  /api/tasks/{taskId}/complete-timed:
    post:
      summary: Complete a timed task
      description: Marks a timed task as completed.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                actualDurationSpentMinutes:
                  type: integer
      responses:
        '200':
          description: Timed task completed
        '400':
          description: Validation error
        '404':
          description: Task not found
        '500':
          description: Internal error
  /api/tasks/{taskId}/complete-bad:
    post:
      summary: Complete a bad task
      description: Marks a bad task as complete.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Bad task completed successfully
        '400':
          description: Validation error
        '404':
          description: Task not found
        '500':
          description: Internal error
  /api/tasks/{taskId}/mark-bad:
    put:
      summary: Mark a task as bad
      description: Marks a good task as bad.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task marked as bad
        '400':
          description: Validation error
        '404':
          description: Task not found
        '500':
          description: Internal error
  /api/tasks/{taskId}:
    delete:
      summary: Delete a task
      description: Deletes a task by its ID.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Task deleted
        '404':
          description: Task not found
        '500':
          description: Internal error
  /api/social-blocker:
    post:
      summary: Setup social blocker
      description: Sets up the social media blocker with the given password, start and end time.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                socialPassword:
                  type: string
                socialEnd:
                  type: string
      responses:
        '200':
          description: Social blocker setup
        '400':
          description: Validation error
        '500':
          description: Internal error
  /api/social-blocker/get:
    get:
      summary: Get social blocker info
      description: Get the details about how much the user has progressed through the social media blocker.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Social blocker info
        '404':
          description: Not found
        '500':
          description: Internal error
  /api/social-blocker/end:
    post:
      summary: End social blocker
      description: When the timeout is over, this endpoint is called to end the social media blocker.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Social blocker ended
        '500':
          description: Internal error
  /api/study-plan/generate:
    post:
      summary: Generate Study Plan
      description: This feature is known as "Study Plan" in the app. Internally, it is named timetable.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                chapters:
                  type: object
                deadline:
                  type: string
                clientDate:
                  type: string
      responses:
        '200':
          description: Timetable generated
        '400':
          description: Validation error
        '500':
          description: Internal error
  /api/study-plan:
    post:
      summary: Save study plan
      description: If the user is fine with the study plan preview, this endpoint saves it.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                subjects:
                  type: object
                chapters:
                  type: object
                deadline:
                  type: string
                timetable:
                  type: array
                  items:
                    type: object
      responses:
        '201':
          description: Study plan saved
        '400':
          description: Validation error
        '500':
          description: Internal error
    get:
      summary: Get study plan
      description: Gets the study plan whenever needed.
      security:
        - bearerAuth: []
      parameters:
        - name: clientDate
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Study plan
        '404':
          description: Not found
        '500':
          description: Internal error
    delete:
      summary: Delete study plan
      description: Permanently deletes the study plan of the user so they can setup a new one.
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Study plan deleted
        '500':
          description: Internal error
  /api/study-plan/tasks/{taskId}/complete:
    post:
      summary: Complete a study plan task
      description: Marks a task from the study plan as complete and gives aura for it.
      security:
        - bearerAuth: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                clientDate:
                  type: string
                dateOfTask:
                  type: string
      responses:
        '200':
          description: Study plan task completed
        '400':
          description: Validation error
        '404':
          description: Not found
        '500':
          description: Internal error
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT